+++ plugins/postcss-ic-unit/src/has-supports-at-rule-ancestor.ts
@@ -8,8 +8,8 @@ export function hasSupportsAtRuleAncestor(node: Node): boolean {
 			continue;
 		}
 
-		if ((parent as AtRule).name === 'supports') {
-			if (/\(font-size: \d+ic\)/.test((parent as AtRule).params)) {
+		if ((parent as AtRule).name.toLowerCase() === 'supports') {
+			if (/\(font-size: \d+ic\)/.test((parent as AtRule).params.toLowerCase())) {
 				return true;
 			}
 		}