+++ plugins/postcss-oklab-function/src/on-css-function.ts
@@ -7,7 +7,7 @@ import { oklabToSRgb } from './convert-oklab-to-srgb';
 import { oklchToSRgb } from './convert-oklch-to-srgb';
 
 export function onCSSFunctionSRgb(node: FunctionNode) {
-	const value = node.value;
+	const value = node.value.toLowerCase();
 	const rawNodes = node.nodes;
 	const relevantNodes = rawNodes.slice().filter((x) => {
 		return x.type !== 'comment' && x.type !== 'space';
@@ -71,7 +71,7 @@ export function onCSSFunctionSRgb(node: FunctionNode) {
 export function onCSSFunctionDisplayP3(node: FunctionNode, decl: Declaration, result: Result, preserve: boolean) {
 	const originalForWarnings = valueParser.stringify(node);
 
-	const value = node.value;
+	const value = node.value.toLowerCase();
 	const rawNodes = node.nodes;
 	const relevantNodes = rawNodes.slice().filter((x) => {
 		return x.type !== 'comment' && x.type !== 'space';
@@ -201,12 +201,14 @@ function isNumericNodeHueLike(node: Node): node is WordNode {
 		return false;
 	}
 
+	const unitAndValueUnit = unitAndValue.unit.toLowerCase();
+
 	return !!unitAndValue.number && (
-		unitAndValue.unit === 'deg' ||
-		unitAndValue.unit === 'grad' ||
-		unitAndValue.unit === 'rad' ||
-		unitAndValue.unit === 'turn' ||
-		unitAndValue.unit === ''
+		unitAndValueUnit === 'deg' ||
+		unitAndValueUnit === 'grad' ||
+		unitAndValueUnit === 'rad' ||
+		unitAndValueUnit === 'turn' ||
+		unitAndValueUnit === ''
 	);
 }
 
@@ -228,11 +230,11 @@ function isNumericNodePercentageOrNumber(node: Node): node is WordNode {
 }
 
 function isCalcNode(node: Node): node is FunctionNode {
-	return node && node.type === 'function' && node.value === 'calc';
+	return node && node.type === 'function' && node.value.toLowerCase() === 'calc';
 }
 
 function isVarNode(node: Node): node is FunctionNode {
-	return node && node.type === 'function' && node.value === 'var';
+	return node && node.type === 'function' && node.value.toLowerCase() === 'var';
 }
 
 function isSlashNode(node: Node): node is DivNode {
@@ -425,7 +427,7 @@ function replaceWith(nodes: Array<Node>, oldNode: Node, newNode: Node) {
 }
 
 function normalizeHueNode(dimension: Dimension) {
-	switch (dimension.unit) {
+	switch (dimension.unit.toLowerCase()) {
 		case 'deg':
 			dimension.unit = '';
 			return;