+++ plugins/postcss-trigonometric-functions/src/index.ts
@@ -33,7 +33,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 		postcssPlugin: 'postcss-trigonometric-functions',
 		Declaration(decl) {
 			const transformations = Transformations.filter(
-				transformation => decl.value.includes(transformation.check),
+				transformation => decl.value.toLowerCase().includes(transformation.check),
 			);
 
 			if (!decl || transformations.length === 0) {
@@ -50,12 +50,14 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 				}
 			});
 
-			if (decl.value !== newDeclaration.value) {
-				if (options.preserve) {
-					decl.cloneBefore({ value: newDeclaration.value });
-				} else {
-					decl.value = newDeclaration.value;
-				}
+			if (decl.value === newDeclaration.value) {
+				return;
+			}
+
+			decl.before( newDeclaration );
+
+			if (!options.preserve) {
+				decl.remove();
 			}
 		},
 	};