+++ plugins/postcss-color-rebeccapurple/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS RebeccaPurple
 
+### Unreleased
+
+- Fix case insensitive matching.
+
 ### 7.1.0 (June 10, 2022)
 
 - Added: Typescript typings