+++ plugins/postcss-normalize-display-values/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Normalize Display Values
 
+### Unreleased
+
+- Fix case insensitive matching.
+
 ### 1.0.0 (January 25, 2022)
 
 - Initial version